refactor(search): Move sort responsibility from SearchTracePage into SearchResults
order-by.ts needed to go one level deeper than the previous move: into
SearchResults/, not SearchTracePage/, since that's the only actual
consumer. Same for sortTraceSummaries, toOrderBy, and fromOrderBy,
previously in model/search.ts -- moved into a new SearchResults/sort.ts
alongside order-by.ts, since neither has consumers outside SearchResults.

The deeper fix: SearchTracePage was sorting traceSummaries itself and
threading sortBy/handleSortChange down as props. SearchResults now reads
sortBy/setSortBy from useSearchResultsStore directly and sorts internally
before rendering SelectSort and TraceTable. SearchTracePage drops sortBy,
handleSortChange, and the sortTraceSummaries call entirely, passing
unsorted traceSummaries straight through.

sortTraces (a fourth function in model/search.ts, not named in review)
stays put -- it has no consumers anywhere in the app, so moving it wasn't
asked for and would have been guessing rather than following the review.

Removed a test in SearchTracePage/index.test.jsx that asserted on the
now-removed sortBy/handleSortChange props, and added integration test
coverage in SearchResults/index.test.jsx that changing the sort dropdown
actually re-sorts rendered results and updates the store.
