+++ packages/jaeger-ui/src/components/SearchTracePage/SearchResults/TraceTable.test.tsx
@@ -6,9 +6,9 @@ import { render, screen, fireEvent, within } from '@testing-library/react';
 import { MemoryRouter } from 'react-router-dom';
 import { vi, describe, it, expect, beforeEach } from 'vitest';
 import TraceTable from './TraceTable';
-import * as orderBy from '../order-by';
-import type { OrderBy } from '../order-by';
-import { toOrderBy, fromOrderBy } from '../../../model/search';
+import * as orderBy from './order-by';
+import type { OrderBy } from './order-by';
+import { toOrderBy, fromOrderBy } from './sort';
 import type { Microseconds } from '../../../types/units';
 import { useSearchResultsStore } from '../store.search-results';
 import { formatDatetime } from '../../../utils/date';
