+++ packages/jaeger-ui/src/components/SearchTracePage/SearchResults/index.tsx
@@ -21,8 +21,9 @@ import LoadingIndicator from '../../common/LoadingIndicator';
 import NewWindowIcon from '../../common/NewWindowIcon';
 import SearchResultsDDG from '../../DeepDependencies/traces';
 import { getTracePageLink } from '../../TracePage/url';
-import * as orderBy from '../order-by';
-import type { OrderBy } from '../order-by';
+import * as orderBy from './order-by';
+import type { OrderBy } from './order-by';
+import { sortTraceSummaries } from './sort';
 import { getPercentageOfDuration } from '../../../utils/date';
 
 import { TraceSummary } from '../../../types/trace-summary';
@@ -31,7 +32,9 @@ import './index.css';
 import { getTargetEmptyOrBlank } from '../../../utils/config/get-target';
 import withRouteProps from '../../../utils/withRouteProps';
 import SearchableSelect from '../../common/SearchableSelect';
-import { useSearchResultsStore } from '../store.search-results';
+import { useSearchResultsStore, sanitizeSortBy } from '../store.search-results';
+import { trackSortByChange } from '../SearchForm.track';
+import { useShallow } from 'zustand/react/shallow';
 
 type SearchResultsProps = {
   addTraceToCohort: (summary: TraceSummary) => void;
@@ -48,8 +51,6 @@ type SearchResultsProps = {
   traceSummaries: TraceSummary[];
   uploadedTraceIDs: ReadonlySet<string>;
   rawTraces: unknown[];
-  sortBy: OrderBy;
-  handleSortChange: (sortBy: OrderBy) => void;
 };
 
 type SelectSortProps = {
@@ -95,18 +96,35 @@ export function UnconnectedSearchResults({
   traceSummaries,
   uploadedTraceIDs,
   rawTraces,
-  sortBy,
-  handleSortChange,
   addTraceToCohort,
   removeTraceFromCohort,
 }: SearchResultsProps) {
   const navigate = useNavigate();
-  const viewMode = useSearchResultsStore(s => s.viewMode);
-  const setViewMode = useSearchResultsStore(s => s.setViewMode);
+  const { viewMode, setViewMode, sortBy, setSortBy } = useSearchResultsStore(
+    useShallow(s => ({
+      viewMode: s.viewMode,
+      setViewMode: s.setViewMode,
+      sortBy: s.sortBy,
+      setSortBy: s.setSortBy,
+    }))
+  );
+
+  const handleSortChange = useCallback(
+    (newSortBy: string) => {
+      setSortBy(newSortBy);
+      trackSortByChange(sanitizeSortBy(newSortBy));
+    },
+    [setSortBy]
+  );
+
+  const sortedTraceSummaries = useMemo(
+    () => sortTraceSummaries(traceSummaries, sortBy),
+    [traceSummaries, sortBy]
+  );
 
   const traceSummaryById = useMemo(
-    () => new Map(traceSummaries.map(summary => [summary.traceID, summary])),
-    [traceSummaries]
+    () => new Map(sortedTraceSummaries.map(summary => [summary.traceID, summary])),
+    [sortedTraceSummaries]
   );
 
   const toggleComparison = useCallback(
@@ -156,9 +174,9 @@ export function UnconnectedSearchResults({
     // When URL has lost search params (e.g. after TopNav navigation to bare /search),
     // fall back to the root service of the first result so DDG can build the graph
     const serviceFromUrl = typeof urlState.service === 'string' ? urlState.service : undefined;
-    const service = serviceFromUrl ?? traceSummaries[0]?.rootServiceName;
+    const service = serviceFromUrl ?? sortedTraceSummaries[0]?.rootServiceName;
     navigate(getUrl({ ...urlState, service, view }));
-  }, [location, navigate, traceSummaries]);
+  }, [location, navigate, sortedTraceSummaries]);
 
   const traceResultsView = queryString.parse(location.search).view !== 'ddg';
 
@@ -198,7 +216,7 @@ export function UnconnectedSearchResults({
         {!hideGraph && traceResultsView && (
           <div className="ub-p3 SearchResults--headerScatterPlot">
             <ScatterPlot
-              data={traceSummaries.map(t => {
+              data={sortedTraceSummaries.map(t => {
                 return {
                   x: t.startTime,
                   y: t.duration,
@@ -217,12 +235,14 @@ export function UnconnectedSearchResults({
         )}
         <div className="SearchResults--headerOverview">
           <h2 className="ub-m0 u-flex-1 SearchResults--resultCount">
-            {traceSummaries.length} Trace{traceSummaries.length > 1 && 's'}
+            {sortedTraceSummaries.length} Trace{sortedTraceSummaries.length > 1 && 's'}
           </h2>
           {traceResultsView && viewMode === 'list' && (
             <SelectSort sortBy={sortBy} handleSortChange={handleSortChange} />
           )}
-          {traceResultsView && <DownloadResults traceSummaries={traceSummaries} rawTraces={rawTraces} />}
+          {traceResultsView && (
+            <DownloadResults traceSummaries={sortedTraceSummaries} rawTraces={rawTraces} />
+          )}
           <AltViewOptions traceResultsView={traceResultsView} onDdgViewClicked={onDdgViewClicked} />
           {traceResultsView && (
             <Radio.Group
@@ -248,13 +268,13 @@ export function UnconnectedSearchResults({
       </div>
       {!traceResultsView && (
         <div className="SearchResults--ddg-container">
-          <SearchResultsDDG location={location} traceIDs={traceSummaries.map(s => s.traceID)} />
+          <SearchResultsDDG location={location} traceIDs={sortedTraceSummaries.map(s => s.traceID)} />
         </div>
       )}
       {traceResultsView && diffSelection}
       {traceResultsView && viewMode === 'table' && (
         <TraceTable
-          traceSummaries={traceSummaries}
+          traceSummaries={sortedTraceSummaries}
           maxTraceDuration={maxTraceDuration}
           getLink={getLink}
           sortBy={sortBy}
@@ -266,7 +286,7 @@ export function UnconnectedSearchResults({
       )}
       {traceResultsView && viewMode === 'list' && (
         <ul className="ub-list-reset">
-          {traceSummaries.map(traceSummary => (
+          {sortedTraceSummaries.map(traceSummary => (
             <li className="ub-my3" key={traceSummary.traceID}>
               <ResultItem
                 durationPercent={getPercentageOfDuration(traceSummary.duration, maxTraceDuration)}
