+++ packages/jaeger-ui/src/components/SearchTracePage/SearchResults/order-by.ts
